Removed %$(EXE_EXT) library dependency detection
This is done since the generic makefile cannot figure out where to
download/clone the relevant external dependencies,
this will be implemented in mpm directly (using a topological sort on
all direct and indirect dependencies).
It should be tried if a plain old list of dependencies is enough for
make and makes DAG algorithm is sufficient enough to resolve the
dependencies on its own.